moved AppConfig, ComponentConfig, Configuration, DependencyNode, and ThemeConfig class from model to bean package
