+++ plugin/src/main/java/org/wso2/carbon/uuf/maven/AppMojo.java
@@ -33,9 +33,9 @@ import org.twdata.maven.mojoexecutor.MojoExecutor;
 import org.wso2.carbon.uuf.maven.exception.ParsingException;
 import org.wso2.carbon.uuf.maven.exception.SerializationException;
 import org.wso2.carbon.uuf.maven.model.Bundle;
-import org.wso2.carbon.uuf.maven.model.ComponentConfig;
-import org.wso2.carbon.uuf.maven.model.Configuration;
-import org.wso2.carbon.uuf.maven.model.DependencyNode;
+import org.wso2.carbon.uuf.maven.bean.ComponentConfig;
+import org.wso2.carbon.uuf.maven.bean.Configuration;
+import org.wso2.carbon.uuf.maven.bean.DependencyNode;
 import org.wso2.carbon.uuf.maven.parser.AppConfigParser;
 import org.wso2.carbon.uuf.maven.parser.ComponentConfigParser;
 import org.wso2.carbon.uuf.maven.parser.DependencyTreeParser;
